Move `reaction_views` into `views`
