+++ src/paper/views/paper_views.py
@@ -28,8 +28,8 @@ from rest_framework.response import Response
 
 from analytics.amplitude import track_event
 from discussion.models import Vote
-from discussion.reaction_views import ReactionViewActionMixin
 from discussion.serializers import VoteSerializer
+from discussion.views import ReactionViewActionMixin
 from hub.permissions import IsModerator
 from paper.exceptions import DOINotFoundError, PaperSerializerError
 from paper.filters import PaperFilter
