+++ src/user/tasks.py
@@ -36,7 +36,7 @@ def handle_spam_user_task(user_id, requestor=None):
         for comment in comments.iterator():
             remove_bounties(comment)
             if requestor:
-                from discussion.reaction_views import censor
+                from discussion.views import censor
 
                 censor(requestor, comment)
                 comment.refresh_related_discussion_count()
